user.c 삭제
헤더 중복 참조 제거위해 user.c의 함수인 LogOn을 screen.c로 옮기고,
user.c 제거
